plugins folder and measurement_base module; #182 review p3
