+++ src/ixdat/__init__.py
@@ -9,14 +9,14 @@ __author__ = "Soren B. Scott, Kenneth Nielsen, Anna Winiwarter, et al"
 __email__ = "sbs@chem.ku.dk"
 __license__ = "MIT"
 
-from .measurements import Measurement
+from .measurement_base import Measurement
 from .spectra import Spectrum
 from . import db
 from . import techniques
 from . import plotters
 from . import exporters
 from . import config
-from .config import plugins
+from .plugins import plugins
 
 # I like this to be sure I'm importing from where I think I am:
 print(f"importing ixdat v{__version__} from {__file__}")
